Koordinaten Int zu Float geändert.

+++ Client/src/main/java/view2/component/world/InternUserAvatar.java
@@ -48,7 +48,7 @@ public class InternUserAvatar extends UserAvatar {
         if (currentDirection != null) {
             Vector2 newPosition = body.getPosition();
             Chati.getInstance().getServerSender().send(ServerSender.SendAction.AVATAR_MOVE,
-                    (int) (newPosition.x * WorldScreen.PPM), (int) (newPosition.y * WorldScreen.PPM));
+                    newPosition.x * WorldScreen.PPM, newPosition.y * WorldScreen.PPM);
         }
     }
